+++ tauri/src/hooks/useXlsxSchema.ts
@@ -1,4 +1,4 @@
-import { useCallback } from "react";
+import { useCallback, useEffect, useState } from "react";
 import type { Dispatch, SetStateAction } from "react";
 import { useEnviroment } from "../context/EnviromentProvider";
 import type { SrcInfo } from "../app/form/FormElementProp";
@@ -132,4 +132,25 @@ export const useXlsxSheets = () => {
 			.then((r) => r.json())
 			.catch(() => []);
 	}, [apiUrl]);
+};
+
+export const useSrcInfoSheets = (srcInfo: SrcInfo | undefined): string[] => {
+	const [sheetNames, setSheetNames] = useState<string[]>([]);
+	const loadSheets = useXlsxSheets();
+	const srcPath = srcInfo?.srcPath ?? "";
+	const regTableInclude = srcInfo?.regTableInclude ?? "";
+	const regTableExclude = srcInfo?.regTableExclude ?? "";
+	const recursive = srcInfo?.recursive ?? "";
+	const regInclude = srcInfo?.regInclude ?? "";
+	const regExclude = srcInfo?.regExclude ?? "";
+	const extension = srcInfo?.extension ?? "";
+
+	useEffect(() => {
+		if (!srcPath) {
+			return;
+		}
+		loadSheets({ srcPath, regTableInclude, regTableExclude, recursive, regInclude, regExclude, extension }).then(setSheetNames);
+	}, [srcPath, regTableInclude, regTableExclude, recursive, regInclude, regExclude, extension, loadSheets]);
+
+	return sheetNames;
 };
